moved snapshot helpers to a separate file
